Link Battle Improvements
Text speed is now set to "fast" when in a link battle.
Forfeiting a link battle will now cause the player to lose, and the opponent to win. If both players forfeit, the battle counts as a draw.
New text for forfeiting.
Confirmation text when selecting "Run".
Remove some unreferenced functions.
The party is healed when entering the Colosseum.