Subway Rewards & Boss Groundwork
Winning a set in the "hard" level group will grant the player more BP.
After a 35-battle streak, the player will be challenged by a boss character with a static team. After defeating the boss, the player will be rewarded a large sum of BP depending on the level group.