Map Script Updates
The text box for returning items to the Lucky Punch girl will no longer progress automatically.
The Day Care man will only stop the player once, and the scene will be reset if he is interacted with from the side.
The player will no longer enter the battle area of the Trainer House facing down.