More Battle Sim Fixes
Add dialogue if the player is unable to gain any points during a Battle Sim run.
Change the move relearner to the correct kind.
Change whiting out in the Battle Sim and Past to no longer flash the screen twice.
Play the warp sound for exiting the Battle Sim prior to the screen fading back in.
Fix collision in the sky world.