Implement Difficulty Options
Add a menu when starting a new game(or inputting the correct button combo on the continue screen) that will allow the player to select a difficulty mode.
Relaxed will remove the obedience check for player-owned Pokemon.
Expert will force the set battle style, disallow items in trainer battles, disallow saving from the start menu(moving the option to PCs), and resetting the game if the player loses a battle.

Remove the NPC that explains obedience when playing on Relaxed mode.
Add an icon to the save screen and trainer card for both Relaxed and Expert mode.
Add a wram constant for each difficulty mode.
Add a font icon to denote Relaxed mode.