Fossil Changes & Related
Change the fossil event so the player manually obtains the Pokemon from a nearby incubator.
Respawn the Sabre Fossil if the player manages to knock out Skelegon.
Add text to the incubator tubes found in the Mewtwo Lab.
Replace the "Trap" battle type with "Force Item," with the latter gaining all of the intricacies of the former.
Change the battle intro text when starting battle with a "Force Item" encounter.