Modify Celebi-Related Events
Fix error where Celebi would not be respawned by the developer battle or Hall of Fame.
Remove "Forest is Restless" flag from the past version of Occident Forest, and set it after the boss battle against Celebi.
Set the "Forest is Restless" flag when Celebi is respawned by a developer battle or Hall of Fame entry.